Improve the Filter Key Track resolution from Q8 to Q16 fixed-point

+++ Digital-Synth-PRA32-U2/pra32-u2-synth.h
@@ -544,7 +544,7 @@ public:
           }
           m_osc.note_on<0>(note_number);
           m_lfo.trigger_lfo();
-          uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+          int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
           m_eg[0].note_on(velocity, osc_pitch_0);
           m_eg[1].note_on(velocity, osc_pitch_0);
         } else {
@@ -578,7 +578,7 @@ public:
         m_osc.set_portamento<0>(m_portamento);
         m_osc.note_on<0>(note_number);
         m_lfo.trigger_lfo();
-        uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+        int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
         m_eg[0].note_on(velocity, osc_pitch_0);
         m_eg[1].note_on(velocity, osc_pitch_0);
       }
@@ -590,7 +590,7 @@ public:
       m_osc.note_on<0>(note_number);
       m_last_note_on_index = 0;
 
-      uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+      int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
       m_eg[0].note_on(velocity, osc_pitch_0);
       m_eg[1].note_on(velocity, osc_pitch_0);
     } else if (m_note_on_number[1] == note_number) {
@@ -602,11 +602,11 @@ public:
       m_last_note_on_index = 1;
 
       if (m_voice_mode == VOICE_POLYPHONIC) {
-        uint16_t osc_pitch_1 = m_osc.get_osc_pitch(1);
+        int32_t osc_pitch_1 = m_osc.get_osc_pitch(1);
         m_eg[2].note_on(velocity, osc_pitch_1);
         m_eg[3].note_on(velocity, osc_pitch_1);
       } else {
-        uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+        int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
         m_eg[0].note_on(velocity, osc_pitch_0);
         m_eg[1].note_on(velocity, osc_pitch_0);
       }
@@ -619,11 +619,11 @@ public:
       m_last_note_on_index = 2;
 
       if (m_voice_mode == VOICE_POLYPHONIC) {
-        uint16_t osc_pitch_2 = m_osc.get_osc_pitch(2);
+        int32_t osc_pitch_2 = m_osc.get_osc_pitch(2);
         m_eg[4].note_on(velocity, osc_pitch_2);
         m_eg[5].note_on(velocity, osc_pitch_2);
       } else {
-        uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+        int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
         m_eg[0].note_on(velocity, osc_pitch_0);
         m_eg[1].note_on(velocity, osc_pitch_0);
       }
@@ -636,11 +636,11 @@ public:
       m_last_note_on_index = 3;
 
       if (m_voice_mode == VOICE_POLYPHONIC) {
-        uint16_t osc_pitch_3 = m_osc.get_osc_pitch(3);
+        int32_t osc_pitch_3 = m_osc.get_osc_pitch(3);
         m_eg[6].note_on(velocity, osc_pitch_3);
         m_eg[7].note_on(velocity, osc_pitch_3);
       } else {
-        uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+        int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
         m_eg[0].note_on(velocity, osc_pitch_0);
         m_eg[1].note_on(velocity, osc_pitch_0);
       }
@@ -769,11 +769,11 @@ public:
       }
 
       if (m_voice_mode == VOICE_POLYPHONIC) {
-        uint16_t osc_pitch = m_osc.get_osc_pitch(note_on_osc_index);
+        int32_t osc_pitch = m_osc.get_osc_pitch(note_on_osc_index);
         m_eg[(note_on_osc_index << 1) + 0].note_on(velocity, osc_pitch);
         m_eg[(note_on_osc_index << 1) + 1].note_on(velocity, osc_pitch);
       } else {
-        uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+        int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
         m_eg[0].note_on(velocity, osc_pitch_0);
         m_eg[1].note_on(velocity, osc_pitch_0);
       }
@@ -823,7 +823,7 @@ public:
 
           if (m_voice_mode == VOICE_MONOPHONIC) {
             m_lfo.trigger_lfo();
-            uint16_t osc_pitch_0 = m_osc.get_osc_pitch(0);
+            int32_t osc_pitch_0 = m_osc.get_osc_pitch(0);
             m_eg[0].note_on(255, osc_pitch_0);
             m_eg[1].note_on(255, osc_pitch_0);
           }
